\docType{data}
\name{Ph1data}
\alias{Ph1data}
\title{The measurement of snowfall in inches for the Phase I process}
\format{A data frame with 82 rows and 4 variables:
\describe{
   \item{Year}{The year of snowfall season}
   \item{Jan}{The snowfall in January}
   \item{Feb}{The snowfall in Feburary}
   \item{Mar}{The snowfall in March}
}}
\usage{
diamonds
}
\description{
A dataset containing the measurement of snowfall in inches for the Phase I process. The variables are as follows:
}
\keyword{datasets}