\docType{data}
\name{Ph2data}
\alias{Ph2data}
\title{The measurement of snowfall in inches for the Phase II process}
\format{A data frame with 51 rows and 4 variables:
\describe{
   \item{Year}{The year of snowfall season}
   \item{Jan}{The snowfall in January}
   \item{Feb}{The snowfall in Feburary}
   \item{Mar}{The snowfall in March}
}}
\description{
A dataset containing the measurement of snowfall in inches for the Phase II process. Note that the measurements between 2000 and 2004 are missing. The variables are as follows:
}
\keyword{datasets}